% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_pip_bar.R
\name{add.pip.bar.hist}
\alias{add.pip.bar.hist}
\title{Add a histogram-style series of pip bars to a plot}
\usage{
add.pip.bar.hist(x, ns, bar.width = 0.5, bar.width.n = 2,
  line.col = c("white", "white"), bar.col = c("red", "blue"), line.wd = 1,
  top.cex = 1, ...)
}
\arguments{
\item{x}{array of x axis values where to center the bars}

\item{ns}{array of frequencies}

\item{bar.width}{The width of a bar in x units}

\item{bar.width.n}{The number of columns of pips to break the bars into}

\item{line.col}{A vector of length 2 giving the colors of the pip lines for the top and bottom bars}

\item{bar.col}{A vector or array giving the colors of the bars (see details)}

\item{line.wd}{Width of the pip lines}

\item{top.cex}{The magnification of the top labels relative to par(cex.axis)}

\item{...}{Unused additional arguments}
}
\value{
Returns NULL invisibly.
}
\description{
Add a histogram-style series of pip bars to a plot
}

