% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diverging_pip_plot.R
\name{diverging_pip_plot}
\alias{diverging_pip_plot}
\title{Create a diverging pip bar plot from frequencies in an array}
\usage{
diverging_pip_plot(ns, bar.width = 0.3, bar.width.n = 2,
  line.col = c("white", "white"), line.wd = 1, bar.col = c("red", "blue"),
  xlab = "", box = TRUE, tick.every = 5, sym = FALSE,
  cluster.width = 0.3, panel.lty = 0, stacked = FALSE, hist = FALSE,
  top.cex = 1)
}
\arguments{
\item{ns}{A 2 by N by M array (see details)}

\item{bar.width}{The width of a bar in x units}

\item{bar.width.n}{The number of columns of pips to break the bars into}

\item{line.col}{A vector of length 2 giving the colors of the pip lines for the top and bottom bars}

\item{line.wd}{Width of the pip lines}

\item{bar.col}{A vector or array giving the colors of the bars (see details)}

\item{xlab}{x axis label}

\item{box}{Create box around entire plot?}

\item{tick.every}{Make a y axis tick after this many pip rows}

\item{sym}{Should the y axis by symmetric around 0?}

\item{cluster.width}{For side-by-side plots, width of all bars together in x axis units}

\item{panel.lty}{The line type for dividing lines across bars/clusters}

\item{stacked}{If TRUE, create a stacked bar plot (see details)}

\item{hist}{If TRUE, create a histogram-style bar chart (see details)}

\item{top.cex}{The magnification of the top labels relative to par(cex.axis)}
}
\value{
Returns NULL invisibly.
}
\description{
Create a diverging pip bar plot from frequencies in an array
}
\examples{
## A basic bar plot
frq = array( c(10,40,20,10,30,20), dim = c(2,4))

rownames(frq) = c("Women","Men")
colnames(frq) = 1995:1998

diverging_pip_plot(frq, bar.width = .3, bar.width.n = 4, sym = TRUE)

## Side-by-side bar plot
frq = array( c(10,40,20,10,
               30,20,15,20,
               15,15,20,22,
               11,23,54,23), dim = c(2,4,2))

rownames(frq) = c("Women","Men")
colnames(frq) = 1995:1998
dimnames(frq)[[3]] = c("a","b")

bar.col = array(c("red","blue","darkred","darkblue"),dim=c(2,2))

diverging_pip_plot(frq, bar.width = .2, bar.width.n = 4, bar.col = bar.col,
                   sym = TRUE, cluster.width = .4, panel.lty = 1)

## Stacked bar plot
diverging_pip_plot(frq, bar.width = .5, bar.width.n = 6, bar.col = bar.col,
                   sym = TRUE, cluster.width = .2, panel.lty = 1, stacked=TRUE)

## Histogram-style plot

bar.col = array(c("pink","lightblue","red","blue","darkred","darkblue"),dim=c(2,3))

frq = array( c(10,40,20,10,
               30,20,15,20,
               15,15,20,22,
               11,23,54,23,
               20,15,20,15,
               15,20,22,31), dim = c(2,4,3))

rownames(frq) = c("Women","Men")
colnames(frq) = 1995:1998
dimnames(frq)[[3]] = c("a","b","c")

diverging_pip_plot(frq, bar.width = .2, bar.width.n = 6, bar.col = bar.col,
                   sym = TRUE, cluster.width = .3, panel.lty = 1, hist=TRUE)
}

