\name{divergingPips-package}
\alias{divergingPips-package}
\alias{divergingPips}
\docType{package}
\title{
\packageTitle{divergingPips}
}
\description{
\packageDescription{divergingPips}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{divergingPips}
\packageIndices{divergingPips}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{divergingPips}

Maintainer: \packageMaintainer{divergingPips}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\examples{
#~~ simple examples of the most important functions ~~
}
